+++ coral-trino/src/main/java/com/linkedin/coral/trino/rel2trino/Calcite2PrestoUDFConverter.java
@@ -3,7 +3,7 @@
  * Licensed under the BSD-2 Clause license.
  * See LICENSE in the project root for license information.
  */
-package com.linkedin.coral.presto.rel2presto;
+package com.linkedin.coral.trino.rel2trino;
 
 import org.apache.calcite.rel.RelNode;
 import org.apache.calcite.rel.RelShuttle;
@@ -34,7 +34,7 @@ import com.linkedin.coral.com.google.common.collect.ImmutableList;
 import com.linkedin.coral.com.google.common.collect.ImmutableMultimap;
 import com.linkedin.coral.com.google.common.collect.Multimap;
 import com.linkedin.coral.hive.hive2rel.functions.GenericProjectFunction;
-import com.linkedin.coral.presto.rel2presto.functions.GenericProjectToPrestoConverter;
+import com.linkedin.coral.trino.rel2trino.functions.GenericProjectToPrestoConverter;
 
 
 public class Calcite2PrestoUDFConverter {
