+++ coral-trino/src/main/java/com/linkedin/coral/trino/rel2trino/CalcitePrestoUDFMap.java
@@ -3,7 +3,7 @@
  * Licensed under the BSD-2 Clause license.
  * See LICENSE in the project root for license information.
  */
-package com.linkedin.coral.presto.rel2presto;
+package com.linkedin.coral.trino.rel2trino;
 
 import java.util.Collection;
 import java.util.HashMap;
@@ -18,9 +18,9 @@ import com.linkedin.coral.com.google.common.collect.ImmutableMultimap;
 import com.linkedin.coral.hive.hive2rel.functions.HiveFunction;
 import com.linkedin.coral.hive.hive2rel.functions.HiveRLikeOperator;
 import com.linkedin.coral.hive.hive2rel.functions.StaticHiveFunctionRegistry;
-import com.linkedin.coral.presto.rel2presto.functions.PrestoElementAtFunction;
+import com.linkedin.coral.trino.rel2trino.functions.PrestoElementAtFunction;
 
-import static com.linkedin.coral.presto.rel2presto.UDFMapUtils.*;
+import static com.linkedin.coral.trino.rel2trino.UDFMapUtils.*;
 
 
 public class CalcitePrestoUDFMap {
