+++ coral-trino/src/main/java/com/linkedin/coral/trino/rel2trino/UDFTransformer.java
@@ -3,7 +3,7 @@
  * Licensed under the BSD-2 Clause license.
  * See LICENSE in the project root for license information.
  */
-package com.linkedin.coral.presto.rel2presto;
+package com.linkedin.coral.trino.rel2trino;
 
 import java.util.ArrayList;
 import java.util.HashMap;
