+++ coral-trino/src/test/java/com/linkedin/coral/trino/rel2trino/HiveToPrestoConverterTest.java
@@ -3,7 +3,7 @@
  * Licensed under the BSD-2 Clause license.
  * See LICENSE in the project root for license information.
  */
-package com.linkedin.coral.presto.rel2presto;
+package com.linkedin.coral.trino.rel2trino;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
